rename to match the file naming of unlock-app
